final/requirements.txt
Django==2.2.3
django-bootstrap4==0.0.8
factory-boy==2.12.0
ipython==7.6.1
celery==4.3.0
redis==3.2.1
psycopg2==2.8.3